Fix kill NodePi child process

+++ extraResources/bashScripts/helpers/package_json_back.sh
@@ -3,6 +3,6 @@
 NODE_PI_FILE_PREFIX=$1
 PACKAGE_JSON_BACKUP_FILE="${NODE_PI_FILE_PREFIX}package.json.bak"
 
-if [ ! -f "${PACKAGE_JSON_BACKUP_FILE}" ]; then
+if [ -f "package.json" ] && [ ! -f "${PACKAGE_JSON_BACKUP_FILE}" ]; then
   cp "package.json" "${PACKAGE_JSON_BACKUP_FILE}"
 fi
